Un-write recovery section
It doesn't make sense to have the main avionics unit the nosecone and blow it away for recovery when we're doing all the work in the main body avionics section.  Will be talking to SD/FD/recovery about moving main avionics bay to main body and remote unit in the nosecone if necessary at all.

Signed-off-by: Misha Turnbull <mishaturnbull@gmail.com>